WMS_AI_PythonApp/WMS_STEP_ONE_Modular_Final_bin_count/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
supabase
python-dotenv
python-multipart
requests
requests-oauthlib
pydantic[email]